chore: deploy/retrieve to/from manifest working
