+++ test/utilities/fileSystem.ts
@@ -50,55 +50,3 @@ export async function createCustomObjects(testSetup: TestSetup): Promise<void> {
     }
   });
 }
-
-export async function createManifestFile(testSetup: TestSetup): Promise<void> {
-  // const projectPath = testSetup.projectFolderPath;
-  // const tempFolderPath = testSetup.tempFolderPath;
-  // const source = path.join(
-  //   tempFolderPath!,
-  //   '..',
-  //   'test',
-  //   'testData',
-  //   'manifest'
-  // );
-  // fs.copy(source, projectPath!, { recursive: true }, async error => {
-  //   if (error) {
-  //     log(`Failed in copying manifest file ${error.message}`);
-  //     await testSetup.tearDown();
-  //     throw error;
-  //   }
-  // });
-  const workbench = await browser.getWorkbench();
-  const editorView = workbench.getEditorView();
-
-  const tempFolderPath = testSetup.tempFolderPath;
-  // Using the Command palette, run File: New File...
-  const inputBox = await runCommandFromCommandPrompt(
-    workbench,
-    'Create: New File...',
-    1
-  );
-
-  // Set the name of the new manifest file
-  await inputBox.setText(path.join(tempFolderPath!, '/manifest/manifest.xml'));
-  await inputBox.confirm();
-  await inputBox.confirm();
-  const textEditor = (await editorView.openEditor(
-    'manifest.xml'
-  )) as TextEditor;
-  const content = [
-    `<?xml version="1.0" encoding="UTF-8"?>`,
-    `<Package xmlns="http://soap.sforce.com/2006/04/metadata">`,
-    `\t<types>`,
-    `\t\t<members>Customer__c</members>`,
-    `\t\t<members>Product__c</members>`,
-    `\t\t<name>CustomObject</name>`,
-    `\t</types>`,
-    `\t<version>57.0</version>`,
-    `</Package>`
-  ].join('\n');
-
-  await textEditor.setText(content);
-  await textEditor.save();
-  await pause(1);
-}
